refactor: Rename workflows to jobs
