+++ src/jobs/rss.rs
@@ -54,7 +54,7 @@ impl Job for RssWorkflow {
         services: impl Services + Send + Sync + 'static,
     ) -> Result<(), human_errors::Error> {
         let config = services.config();
-        super::CronJob::schedule(&config.workflows.rss, services).await
+        CronJob::schedule(&config.workflows.rss, services).await
     }
 
     #[instrument("workflow.rss.handle", skip(self, job, services), fields(job = %job))]
