+++ src/prelude.rs
@@ -3,6 +3,7 @@ pub use crate::config::{Config, Mergeable};
 pub use crate::db::{Cache, KeyValueStore, Queue};
 pub use crate::filter::{Filter, Filterable};
 pub use crate::job::Job;
+pub use crate::jobs::CronJob;
 pub use crate::services::Services;
 pub use crate::web::OAuth2RefreshToken;
 pub use crate::webhooks::WebhookEvent;
